+++ superfields/src/main/java/org/vaadin/miki/superfields/collections/CollectionField.java
@@ -8,6 +8,7 @@ import com.vaadin.flow.component.HasStyle;
 import com.vaadin.flow.component.HasValue;
 import com.vaadin.flow.component.customfield.CustomField;
 import com.vaadin.flow.component.orderedlayout.FlexLayout;
+import com.vaadin.flow.function.SerializablePredicate;
 import com.vaadin.flow.function.SerializableSupplier;
 import com.vaadin.flow.shared.Registration;
 import org.vaadin.miki.markers.HasIndex;
@@ -26,6 +27,7 @@ import java.util.stream.Collectors;
 
 /**
  * All-purpose field for editing and maintaining values that are {@link Collection}s.
+ * Allows filtering individual elements through {@link #setCollectionElementFilter(SerializablePredicate)} (by default filters out all {@code null}s).
  *
  * @param <T> Type of the element in the collection.
  * @param <C> Type of the collection.
@@ -34,7 +36,8 @@ public class CollectionField<T, C extends Collection<T>> extends CustomField<C>
         implements CollectionController, WithIdMixin<CollectionField<T, C>>, HasStyle,
         WithCollectionValueComponentProviderMixin<T, CollectionField<T, C>>,
         WithHelperMixin<CollectionField<T, C>>, WithHelperPositionableMixin<CollectionField<T, C>>,
-        WithValueMixin<AbstractField.ComponentValueChangeEvent<CustomField<C>, C>, C, CollectionField<T, C>> {
+        WithValueMixin<AbstractField.ComponentValueChangeEvent<CustomField<C>, C>, C, CollectionField<T, C>>,
+        WithCollectionElementFilterMixin<T, CollectionField<T, C>> {
 
     /**
      * CSS class name that will be added to the main layout of this component.
@@ -66,6 +69,8 @@ public class CollectionField<T, C extends Collection<T>> extends CustomField<C>
     private CollectionValueComponentProvider<T, ?> collectionValueComponentProvider;
 
     private boolean valueUpdateInProgress = false;
+
+    private SerializablePredicate<T> collectionElementFilter = Objects::nonNull;
 
     /**
      * Creates new field using {@link #DEFAULT_LAYOUT_PROVIDER} as root layout provider.
@@ -157,7 +162,10 @@ public class CollectionField<T, C extends Collection<T>> extends CustomField<C>
 
     @Override
     protected C generateModelValue() {
-        return this.fields.stream().map(HasValue::getValue).collect(Collectors.toCollection(this.emptyCollectionSupplier));
+        return this.fields.stream()
+                .map(HasValue::getValue)
+                .filter(this.getCollectionElementFilter()) // fixes #374
+                .collect(Collectors.toCollection(this.emptyCollectionSupplier));
     }
 
     @Override
@@ -254,8 +262,19 @@ public class CollectionField<T, C extends Collection<T>> extends CustomField<C>
     }
 
     @Override
-    public CollectionValueComponentProvider<T, ?> getCollectionValueComponentProvider() {
-        return collectionValueComponentProvider;
+    @SuppressWarnings("unchecked")
+    public <W extends Component & HasValue<?, T>> CollectionValueComponentProvider<T, W> getCollectionValueComponentProvider() {
+        return (CollectionValueComponentProvider<T, W>) collectionValueComponentProvider;
+    }
+
+    @Override
+    public void setCollectionElementFilter(SerializablePredicate<T> collectionElementFilter) {
+        this.collectionElementFilter = collectionElementFilter == null ? t -> true : collectionElementFilter;
+    }
+
+    @Override
+    public SerializablePredicate<T> getCollectionElementFilter() {
+        return collectionElementFilter;
     }
 
     @Override
@@ -271,6 +290,7 @@ public class CollectionField<T, C extends Collection<T>> extends CustomField<C>
      * @param index Index to get a component at.
      * @return A non-{@code null} component at a given index.
      */
+    @SuppressWarnings("squid:S1452") // only the type is relevant here, could not find a way to narrow down the event type
     final HasValue<?, T> getField(int index) {
         return this.fields.get(index);
     }